Add insertNull attribute to column to actually insert null columns.
This is feature of cassandra called upsert. Inserting without a column
is actually update of columns that are present in insert.
However for some cases you would actually like to insert null value into
column and that by default was not behaviour of spring-data-cassandra.
Now you can choose per property.