Add saveIfNotExist and updateIf to extend ltw tx support
saveIfNotExists is straightforward extension to save.
It will save entity only if it does not exist and will return
LTWTxResult so you can check if it was saved.

If not you can obtain 'offending' record that is a version of record
that is already saved.

The second option is updateIf which will apply only if given set of
conditions is satisfied. This mechanism can be used to perform
optimistic locking (that is updating a record if it had not been updated
by anyone in the mean time).

Again you get the same LTWTxResult.